# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: trendsAF API Documentation
  description: |
    This API provides CRUD operations for trendsAF, allowing users to interact with the endpoints provided. Each endpoint provides typical HTTP methods such as GET, POST, PUT, and DELETE.
  version: 1.0.0
servers:
  - url: http://165.232.43.233:5080
paths:
...
  /auth/login:
    post:
      summary: User Login
      description: Logs a user into the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successfully logged in
  /auth/confirmation:
    post:
      summary: Confirmation
      description: Confirms a user for profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                country:
                  type: string
                company_type:
                  type: string
                company_size:
                  type: string
                start_year:
                  type: string
                annual_revenue:
                  type: string
                company_role:
                  type: string
                province:
                  type: string
                phone:
                  type: string
      responses:
        '200':
          description: Successfully logged in
  /auth/password_reset_request:
    post:
      summary: Request Password Reset
      description: Sends a password reset request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset request sent
  /auth/password_reset/{token}:
    post:
      summary: Reset Password
      description: Resets the user’s password with the provided token.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        '200':
          description: Password reset successful
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []